requirements.txt
fastapi
uvicorn
pytesseract
Pillow
opencv-python-headless
numpy
python-multipart
